+++ Commands/XygeniRunIncrementalScanCommand.cs
@@ -1,0 +1,44 @@
+using System;
+using System.ComponentModel.Design;
+using System.Threading.Tasks;
+using Microsoft.VisualStudio.Shell;
+using Task = System.Threading.Tasks.Task;
+
+namespace vs2026_plugin.Commands
+{
+    internal sealed class XygeniRunIncrementalScanCommand
+    {
+        public const int CommandId = 0x0102;
+        public static readonly Guid CommandSet = new Guid("c7b39864-46c5-43a9-9892-e31d4e0e5621");
+
+        private readonly AsyncPackage package;
+
+        private XygeniRunIncrementalScanCommand(AsyncPackage package, OleMenuCommandService commandService)
+        {
+            this.package = package ?? throw new ArgumentNullException(nameof(package));
+            commandService = commandService ?? throw new ArgumentNullException(nameof(commandService));
+
+            var menuCommandID = new CommandID(CommandSet, CommandId);
+            var menuItem = new MenuCommand(this.Execute, menuCommandID);
+            commandService.AddCommand(menuItem);
+        }
+
+        public static XygeniRunIncrementalScanCommand Instance { get; private set; }
+
+        public static async Task InitializeAsync(AsyncPackage package)
+        {
+            await ThreadHelper.JoinableTaskFactory.SwitchToMainThreadAsync(package.DisposalToken);
+
+            OleMenuCommandService commandService = await package.GetServiceAsync(typeof(IMenuCommandService)) as OleMenuCommandService;
+            Instance = new XygeniRunIncrementalScanCommand(package, commandService);
+        }
+
+        private void Execute(object sender, EventArgs e)
+        {
+            this.package.JoinableTaskFactory.RunAsync(async delegate
+            {
+                await XygeniCommands.RunIncrementalScanAsync();
+            });
+        }
+    }
+}